+++ Python/codecs.c
@@ -95,7 +95,7 @@ PyObject *_PyCodec_Lookup(const char *encoding)
 {
     PyInterpreterState *interp;
     PyObject *result, *args = NULL, *v;
-    int i, len;
+    Py_ssize_t i, len;
 
     if (encoding == NULL) {
 	PyErr_BadArgument();
